Split app into blueprints
