+++ app/customer/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('customer', __name__)
+
+from app.customer import routes, forms
