+++ app/customer/forms.py
@@ -1,0 +1,26 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, TextAreaField, SubmitField
+from wtforms.validators import DataRequired, Email, ValidationError
+from app.models import Customer
+
+class CreateCustomerForm(FlaskForm):
+    name = StringField('Customer name', validators=[DataRequired()])
+    email = StringField('Email', validators=[Email()])
+    company = StringField('Company name')
+    phone = StringField('Phone number')
+    address = TextAreaField()
+    submit = SubmitField('Create Customer')
+
+    def __init__(self, original_email, *args, **kwargs):
+        super(CreateCustomerForm, self).__init__(*args, **kwargs)
+        self.original_email = original_email
+
+    def validate_email(self, email):
+        if email.data != self.original_email:
+            customer = Customer.query.filter_by(email=self.email.data).first()
+            if customer is not None:
+                raise ValidationError('Please use a different email address.')
+
+
+class DeleteForm(FlaskForm):
+    delete = SubmitField('Delete')
