+++ app/main/routes.py
@@ -1,0 +1,60 @@
+from flask import render_template, redirect, url_for, flash, current_app, request, session
+from flask_login import login_required, current_user, login_user, logout_user
+from flask_principal import identity_changed, Identity, AnonymousIdentity
+from werkzeug.urls import url_parse
+from app import db
+from app.main.forms import LoginForm, RegistrationForm
+from app.models import User
+from app.main import bp
+
+
+@bp.route('/')
+@bp.route('/index')
+@login_required
+def index():
+    return render_template('main/index.html', title='Home')
+
+
+@bp.route('/login', methods=['GET', 'POST'])
+def login():
+    if current_user.is_authenticated:
+        return redirect(url_for('main.index'))
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if user is None or not user.check_password(form.password.data):
+            flash('Invalid username or password')
+            return redirect(url_for('main.login'))
+        login_user(user, remember=form.remember_me.data)
+        identity_changed.send(current_app._get_current_object(),
+                                  identity=Identity(user.id))
+        next_page = request.args.get('next')
+        if not next_page or url_parse(next_page).netloc != '':
+            next_page = url_for('main.index')
+        return redirect(next_page)
+    return render_template('main/login.html', title='Sign In', form=form)
+
+
+@bp.route('/logout')
+def logout():
+    logout_user()
+    for key in ('identity.name', 'identity.auth_type'):
+        session.pop(key, None)
+    identity_changed.send(current_app._get_current_object(),
+                          identity=AnonymousIdentity())
+    return redirect(url_for('main.index'))
+
+
+@bp.route('/register', methods=['GET', 'POST'])
+def register():
+    if current_user.is_authenticated:
+        return redirect(url_for('main.index'))
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        user = User(username=form.username.data, email=form.email.data)
+        user.set_password(form.password.data)
+        db.session.add(user)
+        db.session.commit()
+        flash('You are now registered!')
+        return redirect(url_for('main.login'))
+    return render_template('main/register.html', title='Register', form=form)
