+++ app/templates/base.html
@@ -20,20 +20,20 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a class="navbar-brand" href="{{ url_for('index') }}">Work Portal</a>
+                <a class="navbar-brand" href="{{ url_for('main.index') }}">Work Portal</a>
             </div>
             <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
                 <ul class="nav navbar-nav">
-                    <li><a href="{{ url_for('index') }}">Home</a></li>
-                    <li><a href="{{ url_for('customers') }}">Customers</a></li>
+                    <li><a href="{{ url_for('main.index') }}">Home</a></li>
+                    <li><a href="{{ url_for('customer.customers') }}">Customers</a></li>
                     <li><a href="{{ url_for('list_orders') }}">Orders</a></li>
                     <li><a href="{{ url_for('list_controllers') }}">Controllers</a></li>
                 </ul>
                 <ul class="nav navbar-nav navbar-right">
                     {% if current_user.is_anonymous %}
-                        <li><a href="{{ url_for('login') }}">Login</a></li>
+                        <li><a href="{{ url_for('main.login') }}">Login</a></li>
                     {% else %}
-                        <li><a href="{{ url_for('logout') }}">Logout</a></li>
+                        <li><a href="{{ url_for('main.logout') }}">Logout</a></li>
                     {% endif %}
                 </ul>
             </div>
