+++ app/templates/customer/delete.html
