+++ app/templates/customer/detail_customer.html
