+++ app/templates/main/index.html
@@ -8,7 +8,7 @@
             <h1>Hi, {{ current_user.username.capitalize() }}</h1>
         </div>
         <div class="col-lg-offset-4 col-lg-4">
-            <a class="btn btn-block btn-primary" href="{{ url_for('create_customer') }}">Create Customer</a>
+            <a class="btn btn-block btn-primary" href="{{ url_for('customer.create_customer') }}">Create Customer</a>
         </div>
     </div>
     <br>
@@ -19,7 +19,7 @@
             </div>
         </div>
         <div class="col-lg-2">
-            <a class="btn btn-info btn-block" href="{{ url_for('customers') }}">List Customers</a>
+            <a class="btn btn-info btn-block" href="{{ url_for('customer.customers') }}">List Customers</a>
         </div>
     </div>
     </div>
