+++ app/templates/main/login.html
@@ -9,6 +9,6 @@
         </div>
     </div>
     <br>
-    <p>New User? <a href="{{ url_for('register') }}">Click to Register!</a></p>
+    <p>New User? <a href="{{ url_for('main.register') }}">Click to Register!</a></p>
     <p>
 {% endblock %}
